src/main/java/com/example/madancars/CarService.java
package com.example.madancars;

import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

@RequiredArgsConstructor

@Service
public class CarService {

    private final CarRepository carRepository;

    public List<Car> findAll() {
        return carRepository.findAll();
    }

    public Optional<Car> findById(Long id) {
        return carRepository.findById(id);
    }

    public Car save(Car stock) {
        return carRepository.save(stock);
    }

    public void deleteById(Long id) {
        carRepository.deleteById(id);
    }
}
